Adds payment initialization flow
Implements the payment initialization flow for merchants, including subscription checks and redirection to the payment service provider.

This change introduces a new PaymentController in the PaymentServiceProviderSEP project that handles payment initialization requests. It also adds a PaymentComponent to the paymentServiceProvider-fr project that allows users to select a payment method and be redirected to the PSP.

- Implements payment initialization and subscription handling in the backend.
- Creates payment component in the frontend.
- Secure the resources via AuthGuard.
- Adds new DTOs and model mappers.

+++ WebShop/WebShopSEP/src/main/java/com/example/WebShopSEP/controller/EquipmentController.java
@@ -31,7 +31,7 @@ public class EquipmentController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<EquipmentResponseDTO> getEquipmentById(@PathVariable Integer id) {
+    public ResponseEntity<EquipmentResponseDTO> getEquipmentById(@PathVariable Long id) {
         try {
             EquipmentResponseDTO equipment = equipmentService.findById(id);
             return ResponseEntity.ok(equipment);
@@ -49,7 +49,7 @@ public class EquipmentController {
 
     @PreAuthorize("hasRole('AUTHOR')")
     @PutMapping("/{id}")
-    public ResponseEntity<EquipmentResponseDTO> updateEquipment(@PathVariable Integer id, @RequestBody EquipmentUpdateDTO equipmentUpdateDTO) {
+    public ResponseEntity<EquipmentResponseDTO> updateEquipment(@PathVariable Long id, @RequestBody EquipmentUpdateDTO equipmentUpdateDTO) {
         try {
             EquipmentResponseDTO updatedEquipment = equipmentService.update(id, equipmentUpdateDTO);
             return ResponseEntity.ok(updatedEquipment);
@@ -60,7 +60,7 @@ public class EquipmentController {
 
     @PreAuthorize("hasRole('AUTHOR')")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteEquipment(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteEquipment(@PathVariable Long id) {
         try {
             equipmentService.deleteById(id);
             return ResponseEntity.noContent().build();
