+++ WebShop/WebShopSEP/src/main/java/com/example/WebShopSEP/repository/InsuranceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InsuranceRepository extends JpaRepository<Insurance, Integer> {
+public interface InsuranceRepository extends JpaRepository<Insurance, Long> {
 }
